more typescripting. Move inMemoryInstances to own file;
