+++ packages/server/src/utils.ts
@@ -1,6 +1,6 @@
 import { Socket } from "socket.io";
 import { DefaultEventsMap } from "socket.io/dist/typed-events";
-import { InstanceInMemoryData, InstanceUser, UserSlot } from "./models/Instance";
+import { InstanceInMemoryData, InstanceUser, UserSlot } from "./inMemoryInstances";
 
 function random(mn: number, mx: number) {
   return Math.random() * (mx - mn) + mn;
@@ -10,14 +10,14 @@ export const getRandomArrayElement = (arr: any[]) => {
   return arr[Math.floor(random(1, arr.length)) - 1];
 };
 
-export type RoomState = { usedSlots: number; maxSlots: number; users?: InstanceUser[] };
+export type RoomState = { usedSlots: number; maxSlots?: number; users?: InstanceUser[] };
 
 export function assignClientSlot(
   instance: InstanceInMemoryData,
   roomState: RoomState,
   newClient: Socket<DefaultEventsMap, DefaultEventsMap, DefaultEventsMap, any>,
   requestedSlotIndex: number | null
-) {
+): number | false {
   console.log("requested slot index");
   // override requested slot and assign new client id to it
   if (requestedSlotIndex) {
@@ -43,7 +43,7 @@ export function assignClientSlot(
     return requestedSlotIndex;
   }
 
-  if (roomState.usedSlots + 1 > roomState.maxSlots) {
+  if (roomState.usedSlots + 1 > (roomState?.maxSlots ?? 0)) {
     console.log("no available slot for new client", newClient.id);
     return false;
   }
@@ -60,7 +60,7 @@ export function assignClientSlot(
   // console.log({freeSlotsExcludingLastTried})
 
   // pick random free slot
-  const nextFreeSlotIndex = instance?.settings?.randomPick
+  const nextFreeSlotIndex: number = instance?.settings?.randomPick
     ? getRandomArrayElement(freeSlotsExcludingLastTried).slot_index
     : freeSlotsExcludingLastTried?.[0]?.slot_index ?? 0;
 
@@ -92,12 +92,12 @@ export function resetClientSlot(instance: InstanceInMemoryData, client: Socket) 
   });
 }
 
-export function createRoomState(instance: InstanceInMemoryData, clientsInRoom): RoomState {
+export function createRoomState(instance: InstanceInMemoryData, clientsInRoom?: Set<string>): RoomState {
   const numClients = clientsInRoom ? clientsInRoom.size : 0;
 
   return {
     usedSlots: numClients,
-    maxSlots: instance.settings.slots,
+    maxSlots: instance?.settings?.slots,
     users: instance.users,
   };
 }
